flog.remove {futile.logger}	R Documentation

Remove a logger

Description

In the event that you no longer wish to have a logger registered, use this function to remove it. Then any references to this logger will inherit the next available logger in the hierarchy.

Arguments

name
The logger name to use

Usage

# Remove a logger
flog.remove(name)

Author(s)

Brian Lee Yung Rowe

Examples

flog.threshold(ERROR, name='my.logger')
flog.info("Won't print", name='my.logger')
flog.remove('my.logger')
flog.info("Will print", name='my.logger')

[Package futile.logger version 1.4.3 Index]