ftry {futile.logger}	R Documentation

Wrap a try block in futile.logger

Description

This function integrates futile.logger with the error and warning system so problems can be caught both in the standard R warning system, while also being emitted via futile.logger.

Usage

ftry(expr, error = stop, finally = NULL)

Arguments

expr
The expression to evaluate in a try block

error
An error handler

finally
Pass-through to tryCatch finally

Author(s)

Brian Lee Yung Rowe

Examples

ftry(log(-1))

[Package futile.logger version 1.4.3 Index]